coding_assistant/requirements.txt
google-adk
python-dotenv
gunicorn
google-cloud-aiplatform
google-generativeai>=0.3.0
pydantic
flask>=2.0.0
PyGithub>=2.6.1
